Added initial Python code
This is the initial code I've been working on with a small sample training set.  Training set is not included.  This code is generalizable to accept other cybersecurity training data and predict whether a user input would require a human cybersecurity expert's intervention or if it should be dismissed & for what reason.